use janome if mecab is not available
